requirements.txt
Appium-Python-Client==2.11.1
behave==1.2.6
selenium==4.11.2
urllib3==1.26.6
allure-behave==2.13.2